Reorganize code into packages.
