+++ 5calls/app/src/main/java/org/a5calls/android/a5calls/model/DatabaseHelper.java
@@ -1,4 +1,4 @@
-package org.a5calls.android.a5calls;
+package org.a5calls.android.a5calls.model;
 
 import android.content.ContentValues;
 import android.content.Context;
