+++ 5calls/app/src/main/res/layout/activity_location.xml
@@ -15,7 +15,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         android:orientation="vertical"
-        tools:context="org.a5calls.android.a5calls.LocationActivity">
+        tools:context="org.a5calls.android.a5calls.controller.LocationActivity">
 
         <TextView
             android:layout_width="wrap_content"
